pyproject.toml
[project]
name = "bookstore-new"
version = "0.1.0"
description = ""
authors = [
    {name = "marcelo chaves", email = "marceloluiz.icg@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework (>=3.16.0,<4.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "django-extensions (>=4.1,<5.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
factory-boy = "^3.3.3"
pytest-cov = "^6.2.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "book.settings"
python_files = ["test_*.py"]

